requirements.txt
discord.py>=2.0
feedparser
python-dotenv
beautifulsoup4
